requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiohttp-sse-client==0.2.1
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.8.0
args==0.1.0
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
bip-utils==2.9.3
bitarray==3.1.1
blinker==1.8.2
boto3==1.37.13
botocore==1.37.13
bs4==0.0.2
cachetools==5.3.3
cbor2==5.6.5
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
clint==0.5.1
coincurve==21.0.0
crc32c==2.7.1
crcmod==1.7
cryptography==44.0.1
Django==5.1.6
ecdsa==0.19.0
ed25519-blake2b==1.4.1
et-xmlfile==1.1.0
fastapi==0.115.8
filelock==3.18.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.5.0
gigachat==0.1.38
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
inotify_simple==1.3.5
ipaddress==1.0.23
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.3.0.post0
passlib==1.7.4
pathspec==0.12.1
propcache==0.3.0
proto-plus==1.23.0
protobuf==4.25.3
py-sr25519-bindings==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.22.0
pycryptodomex==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
PyNaCl==1.5.0
pyparsing==3.1.2
PySocks==1.7.1
pyTelegramBotAPI==4.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-magic==0.4.27
python-telegram-bot==21.11.1
pytonconnect==0.3.2
pytoniq==0.1.40
pytoniq-core==0.1.41
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.4
S4==0.4.4
selenium==4.29.0
setuptools==76.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.3
starlette==0.45.3
tabulate==0.9.0
telegram==0.0.1
tonapi==0.1.0
toncenter==0.0.5
tonsdk==1.0.15
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
tvm-valuetypes==0.0.12
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
wsproto==1.2.0
WTForms==3.1.2
x25519==0.0.2
yarl==1.18.3
